+++ tests-optional/test_encoder.py
@@ -25,12 +25,9 @@ from urllib.request import urlretrieve
 
 import faiss
 
-from pyserini.encode import TctColBertDocumentEncoder, DprDocumentEncoder, UniCoilDocumentEncoder
 from pyserini.encode._clip import ClipDocumentEncoder
 from pyserini.search.lucene import LuceneImpactSearcher
 
-
-## We need to de-dup wrt tests/test_encoder
 
 class TestEncode(unittest.TestCase):
     @classmethod
@@ -62,30 +59,6 @@ class TestEncode(unittest.TestCase):
         if not pl.Path(path).resolve().is_file():
             raise AssertionError("File does not exist: %s" % str(path))
 
-    def test_dpr_encoder(self):
-        encoder = DprDocumentEncoder('facebook/dpr-ctx_encoder-multiset-base', device='cpu')
-        vectors = encoder.encode(self.texts[:3])
-        self.assertAlmostEqual(vectors[0][0], -0.59793323, places=4)
-        self.assertAlmostEqual(vectors[0][-1], -0.13036962, places=4)
-        self.assertAlmostEqual(vectors[2][0], -0.3044764, places=4)
-        self.assertAlmostEqual(vectors[2][-1], 0.1516793, places=4)
-
-    def test_tct_colbert_encoder(self):
-        encoder = TctColBertDocumentEncoder('castorini/tct_colbert-msmarco', device='cpu')
-        vectors = encoder.encode(self.texts[:3])
-        self.assertAlmostEqual(vectors[0][0], -0.01649557, places=4)
-        self.assertAlmostEqual(vectors[0][-1], -0.05648308, places=4)
-        self.assertAlmostEqual(vectors[2][0], -0.10293338, places=4)
-        self.assertAlmostEqual(vectors[2][-1], 0.05549275, places=4)
-
-    def test_unicoil_encoder(self):
-        encoder = UniCoilDocumentEncoder('castorini/unicoil-msmarco-passage', device='cpu')
-        vectors = encoder.encode(self.texts[:3])
-        self.assertAlmostEqual(vectors[0]['generation'], 2.2441017627716064, places=4)
-        self.assertAlmostEqual(vectors[0]['normal'], 2.4618067741394043, places=4)
-        self.assertAlmostEqual(vectors[2]['rounding'], 3.9474332332611084, places=4)
-        self.assertAlmostEqual(vectors[2]['commercial'], 3.288801670074463, places=4)
-    
     def test_clip_encoder(self):
         encoder = ClipDocumentEncoder('openai/clip-vit-base-patch32', device='cpu')
         vectors = encoder.encode(self.texts[:3])
